@@ -1,9 +1,9 @@
-module.exports = ({ env }) => ({
-  apiToken: {
-    salt: env('API_TOKEN_SALT'),
-  },
+export default ({ env }) => ({
   auth: {
     secret: env('ADMIN_JWT_SECRET'),
+  },
+  apiToken: {
+    salt: env('API_TOKEN_SALT'),
   },
   transfer: {
     token: {
