@@ -1,6 +1,4 @@
-'use strict';
-
-module.exports = {
+export default {
   /**
    * An asynchronous register function that runs before
    * your application is initialized.
